refactor: extract 'Import' from DeckPicker
